\encoding{utf-8}
\name{dane_nadrzedne_rspo}
\alias{dane_nadrzedne_rspo}
\title{Informaje dla placowkek nadrzędnych}
\usage{
dane_nadrzedne_rspo(rspoStr)
}
\arguments{
  \item{rspoStr}{ciąg znakowy z numerem rspo}
}
\value{
Funkcja zwraca listę, której elementy zawierają wektory
ciągów znakowych z danymi o placówce nadrzędnej i/lub
zespołu.
}
\description{
Funkcja zwraca dla określonego numeru rspo dane placówki
nadrzędnej i zespołu, do którego (o ile) należy.
}

