\encoding{utf-8}
\name{dane_placowka}
\alias{dane_placowka}
\title{Informaje dla pojedynczej placowki}
\usage{
dane_placowka(placowka)
}
\arguments{
  \item{placowka}{węzeł html, zawierający wszystkie
  informacje o placówce.}
}
\value{
Funkcja zwraca wektor ciągów znakowych, którego elementy
to poszczególne elementy to informacje o placowce. Nazwy
poszczególnych elementów tego wektora to typy informacji.
}
\description{
Funkcja przetwarza węzeł dla pojedynczej placówki.
Wynikiem jej działania są wszystkie dane dostępne na
stronie RSPO dla podanej w parametrze placówki.
}

