% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kody_wojewodztw.R
\name{kody_wojewodztw}
\alias{kody_wojewodztw}
\title{Kody wojewodztw}
\usage{
kody_wojewodztw()
}
\value{
funkcja zwraca wektor ciągów znakowych
}
\description{
Funkcja zwraca kody województw, używane przez strony RSPO w kodzie html.
}

