% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nadrzedne_rspo.R
\name{nadrzedne_rspo}
\alias{nadrzedne_rspo}
\title{Numer RSPO dla placowek nadrzednych}
\usage{
nadrzedne_rspo(rspoStr)
}
\arguments{
\item{rspoStr}{ciąg znakowy z numerem rspo}
}
\value{
Funkcja zwraca wektor znakowy, który zawiera numery RSPO.
}
\description{
Funkcja zwraca dla określonego numeru RSPO, numery RSPO placówek nadrzędnych
i zespołów, do których należy.
}

