\encoding{utf-8}
\name{polacz_pliki}
\alias{polacz_pliki}
\title{Pliki województw}
\usage{
polacz_pliki(pliki, nazwaPliku)
}
\arguments{
  \item{pliki}{wektor zawierający ścieżki do łączonych
  plików csv.}

  \item{nazwaPliku}{ścieżka do pliku wynikowego}
}
\value{
funkcja nic nie zwraca
}
\description{
Funkcja łączy pliki z danymi placówek do jednego pliku
csv.
}

