\encoding{utf-8}
\name{rspo_dane}
\alias{rspo_dane}
\title{Pobieranie danych z RSPO}
\usage{
rspo_dane(plik)
}
\arguments{
  \item{plik}{Ścieżka do pliku csv.}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Funkcja pobiera dane szkół podstawowych, gimnazjów,
liceów ogólnokształcących oraz techników. Pobrane dane
zapisuje do pliku csv.
}
\examples{
#rspo_dane("rspo.csv")
}

