\encoding{utf-8}
\name{szczegolowe_dane_placowka}
\alias{szczegolowe_dane_placowka}
\title{Szczegółowe informacje dla placówki}
\usage{
szczegolowe_dane_placowka(placowka)
}
\arguments{
  \item{placowka}{węzeł html, zawierający informacje o
  placówce.}
}
\value{
Funkcja zwraca tablicę ciągów znakowych z informacjami o
placówce.
}
\description{
Funkcja przetwarza węzeł dla pojedynczej placówki.
}

