% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/szkoly_woj_strona.R
\name{szkoly_woj_strona}
\alias{szkoly_woj_strona}
\title{Informacje o placowkach}
\usage{
szkoly_woj_strona(kodWoj, numerStrony)
}
\arguments{
\item{kodWoj}{kod województwa (jako wynik funkcji kody_wojewodztw)}

\item{numerStrony}{numer strony wyszukiwania, z której informacje chcemy
pobrać}
}
\value{
funkcja zwraca listę węzłów html z informacjami o placówkach
}
\description{
Funkcja zwraca informacje o placówkach, które znajdują się na jednej stronie
z wynikami wyszukiwania dla określonego wojewodztwa.
Przez placówkę rozumie się: szkoły podstawowe, gimnazja, technika, licea
ogólnokształcące i profilowane oraz szkoły artystyczne: licea plastyczne,
ogólnokształcące szkoły muzyczne I i II stopnia, ogólnokształcące szkoły
sztuk pięknych, ogólnokształcące szkoły baletowe i Poznańską Szkołę Chóralną.
}

