\encoding{utf-8}
\name{adres_do_korespondencji}
\alias{adres_do_korespondencji}
\title{Adres korespondecyjny}
\usage{
adres_do_korespondencji(placowka)
}
\arguments{
  \item{placowka}{węzeł html placówki}
}
\value{
zwraca wektor danych, który zawiera dane adresowe.
}
\description{
Funkcja parsuje węzeł html placówki w celu pobrania
adresu korespondencji.
}

