% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dane_nadrzedne_rspo.R
\name{dane_nadrzedne_rspo}
\alias{dane_nadrzedne_rspo}
\title{Informaje dla placowkek nadrzednych}
\usage{
dane_nadrzedne_rspo(rspoStr)
}
\arguments{
\item{rspoStr}{ciąg znakowy z numerem rspo}
}
\value{
Funkcja zwraca listę, której elementy zawierają wektory ciągów znakowych z danymi o placówce nadrzędnej i/lub zespołu.
}
\description{
Funkcja zwraca dla określonego numeru rspo dane placówki nadrzędnej i zespołu, do którego (o ile) należy.
}

