\encoding{utf-8}
\name{nadrzedne_rspo}
\alias{nadrzedne_rspo}
\title{Numer RSPO dla placowkek nadrzędnych}
\usage{
nadrzedne_rspo(rspoStr)
}
\arguments{
  \item{rspoStr}{ciąg znakowy z numerem rspo}
}
\value{
Funkcja zwraca wektor znakowy, który zawiera numery RSPO.
}
\description{
Funkcja zwraca dla określonego numeru RSPO, numery RSPO
placówek nadrzędnych i zespołów, do których należy.
}

