% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pobierz_dane_jednego_rspo}
\alias{pobierz_dane_jednego_rspo}
\title{Dane dla numeru rspo}
\usage{
pobierz_dane_jednego_rspo(rspoStr)
}
\arguments{
\item{rspoStr}{ciąg znaków zawierający numer RSPO porządanej placówki.}
}
\value{
Funkcja zwraca wektor z informacjami o placówce.
}
\description{
Funkcja pobiera dane dla jednego numeru RSPO.
}

