% Generated by roxygen2 (4.0.0): do not edit by hand
\name{polacz_pliki}
\alias{polacz_pliki}
\title{Pliki wojewodztw}
\usage{
polacz_pliki(pliki, nazwaPliku)
}
\arguments{
\item{pliki}{wektor zawierający ścieżki do łączonych plików csv.}

\item{nazwaPliku}{ścieżka do pliku wynikowego}
}
\value{
funkcja nic nie zwraca
}
\description{
Funkcja łączy pliki z danymi placówek do jednego pliku csv.
}

