\encoding{utf-8}
\name{polacz_tabele}
\alias{polacz_tabele}
\title{Pliki województw}
\usage{
polacz_tabele(wojList)
}
\arguments{
  \item{wojList}{lista z tablicami danych z poszczególnych
  placówek}
}
\value{
Funkcja zwraca tablicę.
}
\description{
Funkcja łączy pliki z danymi placówek do jednego pliku
csv.
}

