% Generated by roxygen2 (4.0.0): do not edit by hand
\name{polacz_tabele}
\alias{polacz_tabele}
\title{Pliki wojewodztw}
\usage{
polacz_tabele(wojList)
}
\arguments{
\item{wojList}{lista z tablicami danych z poszczególnych placówek}
}
\value{
Funkcja zwraca data frame.
}
\description{
Funkcja łączy pliki z danymi placówek w jeden data frame.
}

