% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/szczegolowe_dane_placowka.R
\name{szczegolowe_dane_placowka}
\alias{szczegolowe_dane_placowka}
\title{Szczegolowe informacje dla placowki}
\usage{
szczegolowe_dane_placowka(placowka)
}
\arguments{
\item{placowka}{węzeł html, zawierający informacje o placówce.}
}
\value{
Funkcja zwraca tablicę ciągów znakowych z informacjami o placówce.
}
\description{
Funkcja przetwarza węzeł dla pojedynczej placówki.
}

