\encoding{utf-8}
\name{wartosc}
\alias{wartosc}
\title{Pojedyncza porcja informacji o placowce}
\usage{
wartosc(nInfo)
}
\arguments{
  \item{nInfo}{węzeł html, zawierający informację o
  placówce.}
}
\value{
Funkcja zwraca dwuelementowy wektor ciągów znakowych,
które pierwszy element określa rodzaj informacji, a drugi
zawiera informację.
}
\description{
Funkcja przetwarza węzeł z pojedynczą informacją (jak
adres, numer telefonu, strona www,..).
}

