% Generated by roxygen2 (4.0.0): do not edit by hand
\name{zmienne_globalne}
\alias{zmienne_globalne}
\title{Zmienne globalne}
\usage{
zmienne_globalne(nazwa)
}
\arguments{
\item{nazwa}{nazwa parametru.}
}
\value{
zwraca ciąg znaków
}
\description{
Funkcja zwraca zmienne globalne urzywane w kodzie innych funkcji z pakietu EWDrspo.
}

